Update test scripts for BonnyCI integration
Previously, all additional software was being installed via the Travis
CI job, and not the scripts it runs. For the BonnyCI integration
added in BonnyCI/hoist#87 to work as expected, we are adding the
software installations to the scripts themselves. With this change,
Travis will require Trusty and sudo access.

Related-Issue: BonnyCI/lore#44
Signed-off-by: Matt Langbehn <matthew.langbehn@gmail.com>