Fix setup instructions for teams/write access
These assumed the repository was under BonnyCI. It is a bit more
complicated for other orgs' repositories. We will also need to fix some
docs for admins so they know to make sure anne-bonny is accepted to the
repo and/or orgs owning the repo as part of the setup completion
process.

Signed-off-by: Clint Byrum <clint@fewbar.com>